Remove redundant write to document
Remove `customersCollection.doc(userId).update(payloadToWrite)` after `customersCollection.doc(userId).set(payloadToWrite, { merge: true })` as this would just overwrite the document directly after creating or merging it with the exact same data. This is probably just an accident. If you want to ensure the document gets either created or replaced, use `set()` without the `merge` flag. If you want to ensure the document gets written, wrap it in a transaction. I just stumbled upon this, thanks for the extension.